Separate data from classes and one class per module
